restrict tasks by user, cache connection to db

+++ public/src/TaskGateway.php
@@ -9,12 +9,15 @@ class TaskGateway
         $this->conn = $database->getConnection();
     }
 
-    public function getAll(): array
+    public function getAllForUser(int $userID): array
     {
         $sql = "SELECT *
                 FROM task
+                WHERE user_id = :user_id
                 ORDER BY name";
-        $stmt = $this->conn->query($sql);
+        $stmt = $this->conn->prepare($sql);
+        $stmt->bindValue(":user_id", $userID, PDO::PARAM_INT);
+        $stmt->execute();
 
         $data = [];
 
@@ -26,13 +29,15 @@ class TaskGateway
         return $data;
     }
 
-    public function get(string $id): array | false
+    public function getForUser(int $userID, string $id): array | false
     {
         $sql = "SELECT * 
                 FROM task
-                WHERE id = :id";
+                WHERE id = :id
+                AND user_id = :user_id";
         $stmt = $this->conn->prepare($sql);
         $stmt->bindValue(":id", $id, PDO::PARAM_INT);
+        $stmt->bindValue(":user_id", $userID, PDO::PARAM_INT);
         $stmt->execute();
         $data = $stmt->fetch(PDO::FETCH_ASSOC);
 
@@ -43,10 +48,10 @@ class TaskGateway
         return $data;
     }
 
-    public function create(array $data): string
+    public function createForUser(int $userId, array $data): string
     {
-        $sql = "INSERT INTO task (name, priority, is_completed)
-                VALUES(:name, :priority, :is_completed)";
+        $sql = "INSERT INTO task (name, priority, is_completed, user_id)
+                VALUES(:name, :priority, :is_completed, :user_id)";
 
         $stmt = $this->conn->prepare($sql);
         $stmt->bindValue(":name", $data["name"], PDO::PARAM_STR);
@@ -58,12 +63,13 @@ class TaskGateway
         }
 
         $stmt->bindValue(":is_completed", $data["is_completed"] ?? false, PDO::PARAM_BOOL);
+        $stmt->bindValue(":user_id", $userId, PDO::PARAM_INT);
 
         $stmt->execute();
         return $this->conn->lastInsertId();
     }
 
-    public function update(int $id, array $data): int
+    public function updateForUser(int $userId, int $id, array $data): int
     {
         $fields = [];
 
@@ -95,11 +101,13 @@ class TaskGateway
         } else {
             $sql = "UPDATE task"
                 . " SET " . implode(", ", $sets)
-                . " WHERE id = :id";
+                . " WHERE id = :id"
+                . " AND user_id = :user_id";
 
             $stmt = $this->conn->prepare($sql);
 
             $stmt->bindValue(":id", $id, PDO::PARAM_INT);
+            $stmt->bindValue(":user_id", $userId, PDO::PARAM_INT);
 
             foreach ($fields as $name => $values) {
                 $stmt->bindValue(":$name", $values[0], $values[1]);
@@ -111,13 +119,15 @@ class TaskGateway
         }
     }
 
-    public function delete(string $id): int
+    public function deleteForUser(int $userId, string $id): int
     {
         $sql = "DELETE FROM task
-                WHERE id = :id";
+                WHERE id = :id
+                AND user_id = :user_id";
 
         $stmt = $this->conn->prepare($sql);
         $stmt->bindValue(":id", $id, PDO::PARAM_INT);
+        $stmt->bindValue(":user_id", $userId, PDO::PARAM_INT);
 
         $stmt->execute();
